Migrating orbis_db as more permanent library wrapper in orbis_client
